# api-doc/swagger.yaml
swagger: '2.0'
info:
  description: This is API documentation of Performance Data Visualizer.
  version: 1.0.0
  title: Performance Data Visualizer API
  contact:
    email: ps@stdin.cz
  license:
    name: MIT license
    url: 'https://opensource.org/licenses/MIT'
host: 'localhost:8001'
basePath: /
tags:
  - name: data
    description: Access performance data (mainly generated by JMH) as JSON
schemes:
  - http
paths:
  /tests:
    get:
      tags:
        - data
      operationId: getTests
      summary: List of all available measurements (benchmarks)
      description: >-
        Get list of all measurements which have at least one data version in a
        configured datastore. Each item contains required and optional metadata
        (for example all metadata from JMH output, such as number of iterations
        or benchmarking mode). Optional metadata are served as key-value pairs.
      produces:
        - application/json
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Test'
        '500':
          description: Internal server error
  '/tests/{testId}/revisions':
    get:
      tags:
        - data
      operationId: getRevisions
      summary: List of different measurement versions in specific test
      description: Returns list of different measurements for specific test
      produces:
        - application/json
      parameters:
        - name: testId
          in: path
          description: ID of test (one of `id` fields from `/tests` endpoint)
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              type: string
              description: Unique version identifier
              example: ver-2017-05-19
            example:
              - ver-2017-05-19
              - ver-2017-05-22
              - ver-last
        '404':
          description: Test not found
        '500':
          description: Internal server error
  '/tests/{testId}/revisions/{revisionId}/data':
    get:
      tags:
        - data
      operationId: getData
      summary: Get data for specific test and version
      description: ''
      produces:
        - application/json
      parameters:
        - name: testId
          in: path
          description: ID of test (one of `id` fields from `/tests` endpoint)
          required: true
          type: string
        - name: revisionId
          in: path
          description: >-
            ID of measurement version (one of the values returned from
            `/tests/${testId}/revisions` endpoint)
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Data'
        '404':
          description: Version or test not found
        '500':
          description: Internal server error
definitions:
  Test:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: string
        description: Unique identifier of this test case
      name:
        type: string
        description: 'Name of test case, usually full method name'
        example: cz.stdin.ps.SampleTest.testMethod
      metadata:
        type: object
        description: Optional test metadata (execution parameters, etc.)
        additionalProperties:
          type: string
          description: Key-value data
  Data:
    type: object
    required:
      - units
    properties:
      units:
        type: string
        example: ns/op
        description: Units of values (all have the same)
      data:
        type: array
        description: >-
          Values separated by distinct measurements (for example JMH forks -
          different JVM instances).
        items:
          type: array
          description: >-
            Values separated by iterations (executions of benchmark inside one
            measurements).
          items:
            type: number
            format: double
            description: Actual single measured value
            example: 18268.275199999567
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'